Invoke the malevolent chronology of Zot
The natural tendency for roguelikes is to encourage cautious behavior.
It's over once you die once, after all, so you should take your time.
That's not a bad thing, but to balance things out and discourage
extremely time-consuming tactics (which tend not to be very fun!),
it's good to have a counterpressure encouraging players to move quickly.

Food was intended to be such a clock, but for a variety of reasons, it
never worked as a clock for anything but the most egregious abuses.
(And not even for that for foodless races, of course.) Scores is a
compelling motivation for some players, but not many - most are just
playing to win. How can we make the most viable way to win also the
most fun?

This commit adds a new 'Zot Clock'. It ticks up over time, and jumps
down whenever you enter a new level - roughly 5000 turns of clock per
level. Once you hit turn 14,500 or so on the clock, the malevolent and
unexplained entity Zot strikes, draining you immediately and then again
once every 100 turns or so. At turn 15,000, you die. Each branch tracks
the 'clock' separately.

5000 turns per level allows for a 250,000 turn 3-rune game, which
seems very generous. The effective cap of 15,000 turns 'stored' at any
given time means that there's somewhat less slack than that implies,
but a slow character can usually do levels in under 2500 turns and
inter-level travel in under 500, so I'm hopeful that normal characters
will rarely if ever see any of this system.

We use 'levels seen' here as a proxy for progress because that's
simple to conceptualize and relatively difficult to abuse - it's hard
to "stash" entire levels as a clock reserve in the same way you could
stash "unexplored map tiles" or "weak monsters". The separate clock
per branch allows players to dip into branches and then bail without
being penalized - the per-branch clock is paused while they're away.

The clock is paused while you're in Abyss, since that doesn't fit the
exploration model well and isn't usually that wise to linger in, and
turned off entirely when you get the ORB, since (a) there's a new
clock then and (b) it'd feel awful to die of time-out on the orb run.
It's possible to scum Pan for time, but I mean, if you're scumming
Pan, just win already!

Chei gives you some extra time, in keeping with the theme of the
god. It's possible that races need different clocks based on their
movement speeds, but I'm hopeful that the difference there is fairly
small relative to the slack in the clock.