Add a lua test that exercises monster spells
This attempts to cast every monster spell, both with and without a foe.
The idea is to at least suss out mistakes in conditioning on a foe
without a targeting flag, and other basic issues. There is no guarantee
that all the spells actually get cast -- since they may still be a waste
of time relative to the artificial scenario. Still, most do get cast so
this is a useful exercise. I implemented this as a lua test rather than
a catch2 test because it requires a setup with actual monsters placed
somewhere in order to work.

The commit also fixes a number of issues found while constructing the
test.