More systematically check `foe` in monster casting
I'd actually removed some of these checks, because as long as the
targeting flag is set properly they are not needed. However, as long as
this API is using a monster * here I decided it is still better coding
practice to have them than not, and it is not as if this API isn't
error-prone. (Really, this code needs a larger refactor where spell
implementation type determines whether there is a `foe` parameter at
all.)